Fix PagingSourceTest generation increment
loadStateFlow may drop states since it is conflated if the observer is
lagging behind, so it is possible for PagingSourceTest to miss generation
updates although they did happen since it uses loadStateFlow to count
generation. This fix moves the generation counting into the PagingData
collector directly so we are guaranteed to increment generation when we
get a new one.

Fixes: 185351978
Test: ./gradlew room:integration-tests:room-testapp-kotlin:cAT
Change-Id: Ib25871da5507d974707c9d10f74c35a89c864e4a
(cherry picked from commit 05cf644cec80274fe26e08a39a4d87ddcb06e0ce)